lib/lzmadecode: Allow for 8 byte reads on 64bit
This adds an optimization to lzma decode to also read from the boot
medium in chunks of 8 bytes if that is the general purpose register
length instead of always 4 bytes. It depends on the cache / memory / spi
controller whether this is faster, but it's likely to be either the same
or faster.

TESTED
- google/vilboz: cached boot medium
64bit before - 32bit - 64bit after
load FSP-M: 35,674 - 35,595 - 34,690
load ramstage: 42,134 - 43,378 - 40,882
load FSP-S: 24,954 - 25,496 - 24,368

- foxconn/g41m: uncached boot medium for testing
64bit before - 32bit - 64bit after
load ramstage: 51,164 - 51,872 - 51,894

Reviewed-on: https://review.coreboot.org/c/coreboot/+/70175
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/lib/lzmadecode.h
@@ -22,10 +22,12 @@
 #ifndef __LZMADECODE_H
 #define __LZMADECODE_H
 
+#include <types.h>
+
 typedef unsigned char Byte;
 typedef unsigned short UInt16;
 typedef unsigned int UInt32;
-typedef UInt32 SizeT;
+typedef size_t SizeT;
 
 #define CProb UInt16
